java移植作業
git-svn-id: https://svn.sourceforge.jp/svnroot/cadencii/Cadencii/trunk@684 9d680d4f-cbfa-477e-9905-73dbafb9b024

+++ bocoree/BNumericUpDown.cs
@@ -21,7 +21,7 @@ namespace bocoree.windows.forms {
 
     public class BNumericUpDown : System.Windows.Forms.NumericUpDown {
         public int getIncrement() {
-            return base.Increment;
+            return (int)base.Increment;
         }
 
         public void setIncrement( int value ) {
